final fixes for validations for messages + updating repos to use a new session for each method

+++ backend-api/backend-api/Features/TaskForum/Controller/MessageController.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using backend_api.Database.ProjectRepository;
+using backend_api.Database.ProjectTaskRepository;
 using backend_api.Features.TaskForum.Models;
 using backend_api.Features.TaskForum.Repository;
 using backend_api.Util;
@@ -22,8 +24,6 @@ namespace backend_api.Features.TaskForum.Controller
         [HttpPost("{channelId}")]
         public ActionResult Post(string channelId, [FromBody]Message messageToCreate, [FromHeader]string userId)
         {
-            // TODO: Check valid auth and access
-            
             // Parse channel guid and user guid
             var channelGuid = ValidationUtil.ParseGuid(channelId);
             var userGuid = ValidationUtil.ParseGuid(userId);
@@ -32,7 +32,19 @@ namespace backend_api.Features.TaskForum.Controller
                 return BadRequest(Errors.InvalidGuid);
             }
             
-            // TODO check channel exists
+            // Check channel exists
+            var existError = CheckChannelExists(channelGuid);
+            if (existError != null)
+            {
+                return existError;
+            }
+
+            // Check user is an owner of the project so has permission to access
+            var error = CheckGroupMember(channelGuid, userGuid);
+            if (error != null)
+            {
+                return error;
+            }
             
             // Check valid message
             if (string.IsNullOrEmpty(messageToCreate.Text.Trim()))
@@ -40,14 +52,10 @@ namespace backend_api.Features.TaskForum.Controller
                 return StatusCode(400, Errors.MessageEmpty);
             }
             
+            // Create
             var result = _messageRepository.Add(messageToCreate, userGuid, channelGuid);
             
-            if (result.IsError)
-            {
-                return StatusCode(500, result.ErrorException.Message);
-            }
-            
-            return StatusCode(201, result.ReturnValue);
+            return result.IsError ? StatusCode(500, result.ErrorException.Message) : StatusCode(201, result.ReturnValue);
         }
         
         // READ
@@ -62,18 +70,25 @@ namespace backend_api.Features.TaskForum.Controller
             {
                 return BadRequest(Errors.InvalidGuid);
             }
-            // TODO
-            // Check auth/user - Unauthorised() and Forbidden() if no access to project
-            // Check channel exists - NotFound()
             
+            // Check channel exists
+            var existError = CheckChannelExists(channelGuid);
+            if (existError != null)
+            {
+                return existError;
+            }
+
+            // Check user is an owner of the project so has permission to access
+            var error = CheckGroupMember(channelGuid, userGuid);
+            if (error != null)
+            {
+                return error;
+            }
+
+            // Read
             var result = _messageRepository.GetAllForChannel(channelGuid);
             
-            if (result.IsError)
-            {
-                return StatusCode(500, result.ErrorException.Message);
-            }
-            
-            return StatusCode(200, result.ReturnValue);
+            return result.IsError ? StatusCode(500, result.ErrorException.Message) : StatusCode(200, result.ReturnValue);
         }
         
         // UPDATE
@@ -81,8 +96,6 @@ namespace backend_api.Features.TaskForum.Controller
         [HttpPatch("{messageId}")]
         public ActionResult Patch(string messageId, [FromBody]Message messageToUpdate, [FromHeader]string userId)
         {
-            // TODO: Check User auth and access - Unauthorised() || Forbidden()
-            
             // Parse message guid and user guid
             var messageGuid = ValidationUtil.ParseGuid(messageId);
             var userGuid = ValidationUtil.ParseGuid(userId);
@@ -92,7 +105,20 @@ namespace backend_api.Features.TaskForum.Controller
             }
             messageToUpdate.Guid = messageGuid;
             
-            // TODO: Check message exists
+            // Check message exists
+            var existError = CheckMessageExists(messageGuid);
+            if (existError != null)
+            {
+                return existError;
+            }
+
+            // Check user is an owner of the project so has permission to access
+            var channel = _messageRepository.GetChannel(messageGuid);
+            var error = CheckGroupMember(channel.ReturnValue.Guid, userGuid);
+            if (error != null)
+            {
+                return error;
+            }
             
             // Check valid channel name
             if (string.IsNullOrEmpty(messageToUpdate.Text.Trim()))
@@ -100,25 +126,16 @@ namespace backend_api.Features.TaskForum.Controller
                 return StatusCode(400, Errors.MessageEmpty);
             }
             
+            // Edit
             var result = _messageRepository.Edit(messageToUpdate);
-            
-            if (result.IsError)
-            {
-                return StatusCode(500, result.ErrorException.Message);
-            }
-            
-            return StatusCode(200);
+
+            return result.IsError ? StatusCode(500, result.ErrorException.Message) : StatusCode(200, result.ReturnValue);
         }
         
         // DELETE
         [HttpDelete("{messageId}")]
         public ActionResult Delete(string messageId, [FromHeader]string userId)
         {
-            // TODO
-            // Check auth - Unauthorised()
-            // Check permissions - Forbidden()
-            // Check message exists - NotFound()
-            
             // Parse message guid and user guid
             var messageGuid = ValidationUtil.ParseGuid(messageId);
             var userGuid = ValidationUtil.ParseGuid(userId);
@@ -126,15 +143,75 @@ namespace backend_api.Features.TaskForum.Controller
             {
                 return BadRequest(Errors.InvalidGuid);
             }
+            
+            // Check message exists
+            var existError = CheckMessageExists(messageGuid);
+            if (existError != null)
+            {
+                return existError;
+            }
 
+            // Check user is an owner of the project so has permission to access
+            var channel = _messageRepository.GetChannel(messageGuid);
+            var error = CheckGroupMember(channel.ReturnValue.Guid, userGuid);
+            if (error != null)
+            {
+                return error;
+            }
+
+            // Delete
             var result = _messageRepository.Delete(messageGuid);
             
-            if (result.IsError)
+            return result.IsError ? StatusCode(500, result.ErrorException.Message) : StatusCode(200, result.ReturnValue);
+        }
+        
+        private ActionResult CheckGroupMember(Guid channelGuid, Guid userGuid)
+        {
+            var channelRepository = new ChannelRepository();
+            var task = channelRepository.GetTask(channelGuid);
+            if (task.IsError)
             {
-                return StatusCode(500, result.ErrorException.Message);
+                return StatusCode(500, task.ErrorException.Message);
             }
-            
-            return StatusCode(200);
+
+            var projectTaskRepository = new ProjectTaskRepository();
+            var project = projectTaskRepository.GetProject(task.ReturnValue.Guid);
+            if (project.IsError)
+            {
+                return StatusCode(500, project.ErrorException.Message);
+            }
+
+            var projectRepository = new ProjectRepository();
+            var inGroup = projectRepository.IsGroupMember(userGuid, project.ReturnValue.Guid);
+            if (inGroup.IsError)
+            {
+                return StatusCode(500, inGroup.ErrorException.Message);
+            }
+
+            return !inGroup.ReturnValue ? StatusCode(403, Errors.NotGroupMember) : null;
+        }
+
+        private ActionResult CheckChannelExists(Guid channelGuid)
+        {
+            var channelRepository = new ChannelRepository();
+            var exists = channelRepository.Exists(channelGuid);
+            if (exists.IsError)
+            {
+                return StatusCode(500, exists.ErrorException.Message);
+            }
+
+            return !exists.ReturnValue ? StatusCode(404, Errors.ChannelNotFound) : null;
+        }
+        
+        private ActionResult CheckMessageExists(Guid messageGuid)
+        {
+            var exists = _messageRepository.Exists(messageGuid);
+            if (exists.IsError)
+            {
+                return StatusCode(500, exists.ErrorException.Message);
+            }
+
+            return !exists.ReturnValue ? StatusCode(404, Errors.MessageNotFound) : null;
         }
     }
 }
